silenced warnings after recent python upgrade
In other words, the fangs of the great python strike again.

How on earth do these people think that it's perfectly fine and
dandy to spit out warnings, especially when the script is not
being run in some sort of a "pedantic-all-warnings-on" mode?

This is the last strike and I will rewrite this in something
else the first chance I get, because it's getting absolutely
ridiculous.

If a particular thing worked in a certain way for an extended
period of time, then that thing is not a BUG anymore, but an
actual FEATURE.