added empty direnv configuration file
The great enshitification marches head on into the future.

I wonder how on earth can you mess up something as simple as this.

If the env var is defined, keep using the env var, if the en var is not
defined, load the config and use the format defined there.

The cognitive load and burden must be too much for the vibe-coder-bros
to handle it all, as I do not have any other reasonable explanation.

In the words of Mr. Stevens: Oh, well!

https://www.youtube.com/watch?v=j59tqyuYG70